requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
python-multipart
playwright
beautifulsoup4
httpx